Refactor main component to split responsabilities
- Refactor ProductList component to be responsible only for the products table
- Create new ProductViewer component to hold the app's state and manage the products list and filter tab
- Create FilterTab component to be responsible for the filter options
- Update README with usage instructions
- Refactor styles into correct modular files
- Add "filterConditions" variable and typing, to hold the filters selected by the user
- Add "hiddenColumns" variable to hold the columns to hide
- Add area above products list where hidden columns will be available to click to be shown again
