+++ src/components/FilterTab/FilterTab.module.scss
@@ -1,0 +1,13 @@
+.filters_container {
+    display: flex;
+    flex: 1;
+    border-right: .2rem solid var(--app-text-support-color);
+    max-width: 25%;
+    min-width: 20%;
+    padding: .75rem;
+
+    h5 {
+        margin: 0;
+        font-size: 1rem;
+    }
+}
