+++ src/components/ProductList/ProductList.module.scss
@@ -1,65 +1,30 @@
-main.main_container {
+.productslist_container {
     height: 100%;
-    margin: 1% 3%;
-    border: .1rem solid var(--app-text-support-color);
-    border-radius: 1.5rem;
-    padding: 1rem;
+    width: 100%;
     display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    // gap: 1rem;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-evenly;
+    gap: .5rem;
+    border: .01rem solid red;
 
-    .filters_container {
+    .hidden_columns_container {
+        align-self: flex-start;
         display: flex;
-        flex: 1;
-        border-right: .2rem solid var(--app-text-support-color);
-        max-width: 25%;
-        min-width: 20%;
-        padding: .75rem;
-
-        h5 {
-            margin: 0;
-            font-size: 1rem;
-        }
+        flex-direction: row;
+        
     }
 
-    .products_container {
-        flex: 4;
-        display: flex;
-        flex-direction: column;
-        // justify-content: space-around;
-        justify-content: flex-start;
-        gap: 1.5rem;
-        padding: 1rem;
-        align-items: center;
+    .productslist_wrapper {
+        min-height: 80%;
+        border: .1rem solid var(--app-text-support-color);
+        border-radius: 1.5rem;
+        width: 100%;
+        overflow-y: auto;
+    }
 
-        h3 {
-            font-size: 1.65rem;
-            margin: 0;
-        }
-
-        .buttons_container {
-            align-self: flex-end;
-            display: flex;
-            flex-direction: row;
-            gap: 1rem;
-
-            .file_input_button {
-                display: none;
-            }
-
-        }
-
-        .productslist_container {
-            min-height: 60%;
-            border: .1rem solid var(--app-text-support-color);
-            border-radius: 1.5rem;
-            width: 100%;
-            overflow-y: auto;
-        }
-
-        .pagination_container {
-            background-color: grey;
-        }
+    .pagination_container {
+        text-align: center;
+        background-color: grey;
     }
 }
