+++ src/components/ProductViewer/ProductViewer.module.scss
@@ -1,0 +1,40 @@
+// Product Viewer main content
+main.main_container {
+    height: 100%;
+    margin: 1% 3%;
+    border: .1rem solid var(--app-text-support-color);
+    border-radius: 1.5rem;
+    padding: 1rem;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    // gap: 1rem;
+
+    .products_container {
+        flex: 4;
+        display: flex;
+        flex-direction: column;
+        // justify-content: space-around;
+        justify-content: flex-start;
+        gap: .8rem;
+        padding: .75rem 1rem;
+        align-items: center;
+
+        h3 {
+            font-size: 1.65rem;
+            margin: 0;
+        }
+
+        .buttons_container {
+            align-self: flex-end;
+            display: flex;
+            flex-direction: row;
+            gap: 1rem;
+
+            .file_input_button {
+                display: none;
+            }
+
+        }
+    }
+}
